+++ packages/core/src/js/wrapper.ts
@@ -394,13 +394,13 @@ export const NATIVE: SentryNativeWrapper = {
     let userKeys = null;
     let userDataKeys = null;
     if (user) {
-      const { id, ip_address, email, username, ...otherKeys } = user;
-      // TODO: Update native impl to use geo
-      const requiredUser: Omit<RequiredKeysUser, 'geo'> = {
+      const { id, ip_address, email, username, geo, ...otherKeys } = user;
+      const requiredUser: RequiredKeysUser = {
         id,
         ip_address,
         email,
         username,
+        geo,
       };
       userKeys = this._serializeObject(requiredUser);
       userDataKeys = this._serializeObject(otherKeys);